+++ miniapp/src/test/java/com/rakuten/tech/mobile/miniapp/RealMiniAppSpec.kt
@@ -6,7 +6,12 @@ import com.nhaarman.mockitokotlin2.verify
 import com.rakuten.tech.mobile.miniapp.api.ApiClient
 import com.rakuten.tech.mobile.miniapp.api.ApiClientRepository
 import com.rakuten.tech.mobile.miniapp.display.Displayer
+import com.rakuten.tech.mobile.miniapp.permission.MiniAppCustomPermissionCache
 import com.rakuten.tech.mobile.miniapp.js.MiniAppMessageBridge
+import com.rakuten.tech.mobile.miniapp.permission.MiniAppCustomPermission
+import com.rakuten.tech.mobile.miniapp.permission.MiniAppCustomPermissionResult
+import com.rakuten.tech.mobile.miniapp.permission.MiniAppCustomPermissionType
+import com.rakuten.tech.mobile.miniapp.navigator.MiniAppNavigator
 import com.rakuten.tech.mobile.sdkutils.AppInfo
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.test.runBlockingTest
@@ -24,12 +29,19 @@ class RealMiniAppSpec {
     private val apiClientRepository: ApiClientRepository = mock()
     private val displayer: Displayer = mock()
     private val miniAppDownloader: MiniAppDownloader = mock()
-    private val miniAppInfo = MiniAppInfo(TEST_MA_ID, "", "", Version("", TEST_MA_VERSION_ID))
     private val miniAppInfoFetcher: MiniAppInfoFetcher = mock()
     private val miniAppSdkConfig: MiniAppSdkConfig = mock()
+    private val miniAppCustomPermissionCache: MiniAppCustomPermissionCache = mock()
     private val realMiniApp =
-        RealMiniApp(apiClientRepository, miniAppDownloader, displayer, miniAppInfoFetcher)
+        RealMiniApp(
+            apiClientRepository,
+            miniAppDownloader,
+            displayer,
+            miniAppInfoFetcher,
+            miniAppCustomPermissionCache
+        )
     private val miniAppMessageBridge: MiniAppMessageBridge = mock()
+    private val miniAppNavigator: MiniAppNavigator = mock()
 
     @Before
     fun setup() {
@@ -62,7 +74,21 @@ class RealMiniAppSpec {
 
             verify(miniAppDownloader, times(1)).getMiniApp(TEST_MA_ID)
             verify(displayer, times(1))
-                .createMiniAppDisplay(getMiniAppResult.first, getMiniAppResult.second, miniAppMessageBridge)
+                .createMiniAppDisplay(getMiniAppResult.first, getMiniAppResult.second,
+                    miniAppMessageBridge, null)
+        }
+
+    @Test
+    fun `should create mini app display with correct passing external navigator`() =
+        runBlockingTest {
+            val getMiniAppResult = Pair(TEST_BASE_PATH, TEST_MA)
+            When calling miniAppDownloader.getMiniApp(TEST_MA_ID) itReturns getMiniAppResult
+            realMiniApp.create(TEST_MA_ID, miniAppMessageBridge, miniAppNavigator)
+
+            verify(miniAppDownloader, times(1)).getMiniApp(TEST_MA_ID)
+            verify(displayer, times(1))
+                .createMiniAppDisplay(getMiniAppResult.first, getMiniAppResult.second,
+                    miniAppMessageBridge, miniAppNavigator)
         }
 
     @Test
@@ -106,5 +132,26 @@ class RealMiniAppSpec {
         miniApp.updateConfiguration(miniAppSdkConfig)
 
         verify(miniApp, times(1)).createApiClient(miniAppSdkConfig)
+    }
+
+    @Test
+    fun `should invoke readPermissions from cache when getCustomPermissions is calling`() {
+        val miniAppId = "miniAppId"
+        realMiniApp.getCustomPermissions(miniAppId)
+
+        verify(miniAppCustomPermissionCache).readPermissions(miniAppId)
+    }
+
+    @Test
+    fun `should invoke storePermissions from cache when setCustomPermissions is calling`() {
+        val miniAppCustomPermission = MiniAppCustomPermission(
+            "dummyMiniAppId",
+            listOf(
+                Pair(MiniAppCustomPermissionType.USER_NAME, MiniAppCustomPermissionResult.DENIED)
+            )
+        )
+        realMiniApp.setCustomPermissions(miniAppCustomPermission)
+
+        verify(miniAppCustomPermissionCache).storePermissions(miniAppCustomPermission)
     }
 }