+++ miniapp/src/test/java/com/rakuten/tech/mobile/miniapp/display/DisplayerSpec.kt
@@ -15,7 +15,7 @@ import org.junit.Test
 import org.junit.runner.RunWith
 
 @RunWith(AndroidJUnit4::class)
-class DisplayerTest {
+class DisplayerSpec {
 
     private lateinit var context: Context
     private val miniAppMessageBridge: MiniAppMessageBridge = mock()
@@ -41,6 +41,7 @@ class DisplayerTest {
         Displayer(context, TEST_HA_NAME).createMiniAppDisplay(
             basePath = context.filesDir.path,
             miniAppInfo = TEST_MA,
-            miniAppMessageBridge = miniAppMessageBridge
+            miniAppMessageBridge = miniAppMessageBridge,
+            miniAppNavigator = mock()
         )
 }