+++ miniapp/src/test/java/com/rakuten/tech/mobile/miniapp/display/MiniAppWebviewSpec.kt
@@ -1,26 +1,26 @@
 package com.rakuten.tech.mobile.miniapp.display
 
 import android.content.Context
+import android.content.Intent
 import android.net.Uri
 import android.util.AndroidRuntimeException
+import android.view.View
 import android.view.ViewGroup
 import android.webkit.GeolocationPermissions
 import android.webkit.WebResourceRequest
 import android.webkit.WebResourceResponse
 import androidx.core.net.toUri
+import androidx.test.core.app.ActivityScenario
 import androidx.test.core.app.ApplicationProvider.getApplicationContext
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.webkit.WebViewAssetLoader
-import com.nhaarman.mockitokotlin2.atLeastOnce
+import com.nhaarman.mockitokotlin2.*
 import com.nhaarman.mockitokotlin2.mock
-import com.nhaarman.mockitokotlin2.times
-import com.nhaarman.mockitokotlin2.verify
 import com.rakuten.tech.mobile.miniapp.*
-import com.rakuten.tech.mobile.miniapp.TEST_BODY_CONTENT
-import com.rakuten.tech.mobile.miniapp.TEST_HA_NAME
-import com.rakuten.tech.mobile.miniapp.TEST_URL_HTTPS_1
-import com.rakuten.tech.mobile.miniapp.TEST_URL_HTTPS_2
 import com.rakuten.tech.mobile.miniapp.js.MiniAppMessageBridge
+import com.rakuten.tech.mobile.miniapp.navigator.ExternalResultHandler
+import com.rakuten.tech.mobile.miniapp.navigator.MiniAppExternalUrlLoader
+import com.rakuten.tech.mobile.miniapp.navigator.MiniAppNavigator
 import org.amshove.kluent.*
 import org.junit.Before
 import org.junit.Test
@@ -29,12 +29,13 @@ import org.mockito.Mockito
 import java.io.ByteArrayInputStream
 import kotlin.test.assertTrue
 
-open class BaseWebViewTest {
+open class BaseWebViewSpec {
     lateinit var context: Context
     lateinit var basePath: String
     internal lateinit var miniAppWebView: MiniAppWebView
     lateinit var webResourceRequest: WebResourceRequest
     val miniAppMessageBridge: MiniAppMessageBridge = mock()
+    val miniAppNavigator: MiniAppNavigator = mock()
     internal lateinit var webChromeClient: MiniAppWebChromeClient
 
     @Before
@@ -48,6 +49,7 @@ open class BaseWebViewTest {
             basePath = basePath,
             miniAppInfo = TEST_MA,
             miniAppMessageBridge = miniAppMessageBridge,
+            miniAppNavigator = miniAppNavigator,
             hostAppUserAgentInfo = TEST_HA_NAME,
             miniAppWebChromeClient = webChromeClient
         )
@@ -56,7 +58,7 @@ open class BaseWebViewTest {
 }
 
 @RunWith(AndroidJUnit4::class)
-class MiniAppWebviewTest : BaseWebViewTest() {
+class MiniAppWebviewSpec : BaseWebViewSpec() {
 
     @Test
     fun `for a given app id, creates corresponding view for the caller`() {
@@ -107,6 +109,7 @@ class MiniAppWebviewTest : BaseWebViewTest() {
             basePath = basePath,
             miniAppInfo = TEST_MA,
             miniAppMessageBridge = miniAppMessageBridge,
+            miniAppNavigator = miniAppNavigator,
             hostAppUserAgentInfo = "",
             miniAppWebChromeClient = webChromeClient
         )
@@ -126,9 +129,10 @@ class MiniAppWebviewTest : BaseWebViewTest() {
     @Test
     fun `each mini app should have different domain`() {
         val miniAppWebViewForMiniapp1 = MiniAppWebView(
-            context, miniAppWebView.basePath, TEST_MA, miniAppMessageBridge, TEST_HA_NAME)
+            context, miniAppWebView.basePath, TEST_MA, miniAppMessageBridge, miniAppNavigator, TEST_HA_NAME)
         val miniAppWebViewForMiniapp2 = MiniAppWebView(
-            context, miniAppWebView.basePath, TEST_MA.copy(id = "app-id-2"), miniAppMessageBridge, TEST_HA_NAME)
+            context, miniAppWebView.basePath, TEST_MA.copy(id = "app-id-2"), miniAppMessageBridge,
+            miniAppNavigator, TEST_HA_NAME)
         miniAppWebViewForMiniapp1.url shouldNotBeEqualTo miniAppWebViewForMiniapp2.url
     }
 
@@ -136,10 +140,27 @@ class MiniAppWebviewTest : BaseWebViewTest() {
     fun `MiniAppMessageBridge should be connected with RealMiniAppDisplay`() {
         verify(miniAppMessageBridge, atLeastOnce()).setWebViewListener(miniAppWebView)
     }
+
+    @Test
+    fun `should load url which is internal mini app scheme from external emission`() {
+        miniAppWebView.externalResultHandler.emitResult(TEST_URL_HTTPS_1)
+        miniAppWebView.externalResultHandler.emitResult("mscheme.${miniAppWebView.miniAppInfo.id}://index.html")
+        val intent = Intent().apply {
+            putExtra(MiniAppExternalUrlLoader.returnUrlTag,
+                "https://mscheme.${miniAppWebView.miniAppInfo.id}/miniapp/index.html")
+        }
+        miniAppWebView.externalResultHandler.emitResult(intent)
+
+        miniAppWebView.getLoadUrl() shouldBeEqualTo
+                "https://mscheme.${miniAppWebView.miniAppInfo.id}/miniapp/index.html"
+    }
 }
 
+@Suppress("SwallowedException")
 @RunWith(AndroidJUnit4::class)
-class MiniAppWebClientTest : BaseWebViewTest() {
+class MiniAppWebClientSpec : BaseWebViewSpec() {
+    private val externalResultHandler: ExternalResultHandler = spy()
+    private val miniAppScheme = MiniAppScheme(TEST_MA_ID)
 
     @Test
     fun `for a WebViewClient, it should be MiniAppWebViewClient`() {
@@ -150,8 +171,8 @@ class MiniAppWebClientTest : BaseWebViewTest() {
     fun `should intercept request with WebViewAssetLoader`() {
         val webAssetLoader: WebViewAssetLoader =
             Mockito.spy((miniAppWebView.webViewClient as MiniAppWebViewClient).loader)
-        val webViewClient = MiniAppWebViewClient(context, webAssetLoader,
-            "custom_domain", "custom_scheme")
+        val webViewClient = MiniAppWebViewClient(context, webAssetLoader, miniAppNavigator,
+            externalResultHandler, miniAppScheme)
         val webResourceRequest = getWebResReq(TEST_URL_HTTPS_1.toUri())
 
         webViewClient.shouldInterceptRequest(miniAppWebView, webResourceRequest)
@@ -164,8 +185,8 @@ class MiniAppWebClientTest : BaseWebViewTest() {
     fun `should redirect to custom domain when only loading with custom scheme`() {
         val webAssetLoader: WebViewAssetLoader = (miniAppWebView.webViewClient as MiniAppWebViewClient).loader
         val customDomain = "https://mscheme.${miniAppWebView.miniAppInfo.id}/"
-        val webViewClient = Mockito.spy(MiniAppWebViewClient(context, webAssetLoader, customDomain,
-            "mscheme.${miniAppWebView.miniAppInfo.id}://"))
+        val webViewClient = Mockito.spy(MiniAppWebViewClient(context, webAssetLoader, miniAppNavigator,
+            externalResultHandler, miniAppScheme))
 
         val displayer = Mockito.spy(miniAppWebView)
 
@@ -179,12 +200,11 @@ class MiniAppWebClientTest : BaseWebViewTest() {
     @Test
     fun `should open phone dialer when there is telephone scheme`() {
         val webAssetLoader: WebViewAssetLoader = (miniAppWebView.webViewClient as MiniAppWebViewClient).loader
-        val webViewClient = Mockito.spy(MiniAppWebViewClient(context, webAssetLoader, "",
-            "mscheme.${miniAppWebView.miniAppInfo.id}://"))
+        val webViewClient = Mockito.spy(MiniAppWebViewClient(context, webAssetLoader, miniAppNavigator,
+            externalResultHandler, miniAppScheme))
         val displayer = Mockito.spy(miniAppWebView)
         val phoneUri = "tel:123456"
 
-        @Suppress("SwallowedException")
         try {
             webViewClient.shouldOverrideUrlLoading(displayer, webResourceRequest)
             webViewClient.shouldOverrideUrlLoading(displayer, getWebResReq(phoneUri.toUri()))
@@ -193,6 +213,40 @@ class MiniAppWebClientTest : BaseWebViewTest() {
         }
 
         verify(webViewClient, times(1)).openPhoneDialer(phoneUri)
+    }
+
+    @Test
+    fun `should not open external url loader when there is no config for navigation`() {
+        val webAssetLoader: WebViewAssetLoader = (miniAppWebView.webViewClient as MiniAppWebViewClient).loader
+        val webViewClient = Mockito.spy(MiniAppWebViewClient(context, webAssetLoader, null,
+            externalResultHandler, miniAppScheme))
+        val displayer = Mockito.spy(miniAppWebView)
+        val externalUri = TEST_URL_HTTPS_1.toUri()
+
+        try {
+            webViewClient.shouldOverrideUrlLoading(displayer, getWebResReq(externalUri))
+        } catch (e: AndroidRuntimeException) {
+            // context here is not activity
+        }
+
+        verify(miniAppNavigator, times(0)).openExternalUrl(TEST_URL_HTTPS_1, externalResultHandler)
+    }
+
+    @Test
+    fun `should open external url loader when there is the config for navigation`() {
+        val webAssetLoader: WebViewAssetLoader = (miniAppWebView.webViewClient as MiniAppWebViewClient).loader
+        val webViewClient = Mockito.spy(MiniAppWebViewClient(context, webAssetLoader, miniAppNavigator,
+            externalResultHandler, miniAppScheme))
+        val displayer = Mockito.spy(miniAppWebView)
+        val externalUri = TEST_URL_HTTPS_1.toUri()
+
+        try {
+            webViewClient.shouldOverrideUrlLoading(displayer, getWebResReq(externalUri))
+        } catch (e: AndroidRuntimeException) {
+            // context here is not activity
+        }
+
+        verify(miniAppNavigator).openExternalUrl(TEST_URL_HTTPS_1, externalResultHandler)
     }
 
     @Test
@@ -223,7 +277,7 @@ class MiniAppWebClientTest : BaseWebViewTest() {
 }
 
 @RunWith(AndroidJUnit4::class)
-class MiniAppWebChromeTest : BaseWebViewTest() {
+class MiniAppWebChromeTest : BaseWebViewSpec() {
 
     @Test
     fun `for a WebChromeClient, it should be MiniAppWebChromeClient`() {
@@ -269,6 +323,28 @@ class MiniAppWebChromeTest : BaseWebViewTest() {
             miniAppWebView, TEST_URL_HTTPS_2, TEST_BODY_CONTENT, TEST_VALUE, mock()) shouldBe true
         webChromeClient.onJsPrompt(
             miniAppWebView, TEST_URL_HTTPS_2, TEST_BODY_CONTENT, null, mock()) shouldBe true
+    }
+
+    @Test
+    fun `should only close custom view when exit`() {
+        webChromeClient = Mockito.spy(MiniAppWebChromeClient(context, TEST_MA))
+        webChromeClient.onShowCustomView(mock(), mock())
+        webChromeClient.customView = mock()
+        webChromeClient.onShowCustomView(mock(), mock())
+
+        verify(webChromeClient).onHideCustomView()
+    }
+
+    @Test
+    fun `should execute custom view flow without error`() {
+        ActivityScenario.launch(TestActivity::class.java).onActivity { activity ->
+            val webChromeClient = Mockito.spy(MiniAppWebChromeClient(activity, TEST_MA))
+
+            webChromeClient.onShowCustomView(View(activity), mock())
+            webChromeClient.updateControls()
+            webChromeClient.onHideCustomView()
+            webChromeClient.updateControls()
+        }
     }
 }