+++ testapp/src/main/java/com/rakuten/tech/mobile/testapp/ui/display/MiniAppDisplayActivity.kt
@@ -1,5 +1,6 @@
 package com.rakuten.tech.mobile.testapp.ui.display
 
+import android.app.Activity
 import android.content.Context
 import android.content.Intent
 import android.content.pm.ApplicationInfo
@@ -13,11 +14,15 @@ import androidx.core.app.ActivityCompat
 import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModelProvider
 import com.rakuten.tech.mobile.miniapp.MiniAppInfo
+import com.rakuten.tech.mobile.miniapp.permission.MiniAppCustomPermissionType
 import com.rakuten.tech.mobile.miniapp.js.MiniAppMessageBridge
-import com.rakuten.tech.mobile.miniapp.js.MiniAppPermissionType
+import com.rakuten.tech.mobile.miniapp.permission.MiniAppPermissionType
+import com.rakuten.tech.mobile.miniapp.navigator.MiniAppNavigator
+import com.rakuten.tech.mobile.miniapp.navigator.ExternalResultHandler
 import com.rakuten.tech.mobile.miniapp.testapp.R
 import com.rakuten.tech.mobile.testapp.helper.AppPermission
 import com.rakuten.tech.mobile.testapp.ui.base.BaseActivity
+import com.rakuten.tech.mobile.testapp.ui.permission.CustomPermissionPresenter
 import com.rakuten.tech.mobile.testapp.ui.settings.AppSettings
 import kotlinx.android.synthetic.main.mini_app_display_activity.*
 
@@ -25,7 +30,11 @@ class MiniAppDisplayActivity : BaseActivity() {
 
     private lateinit var appId: String
     private lateinit var miniAppMessageBridge: MiniAppMessageBridge
+    private lateinit var miniAppNavigator: MiniAppNavigator
     private var miniappPermissionCallback: (isGranted: Boolean) -> Unit = {}
+    private lateinit var sampleWebViewExternalResultHandler: ExternalResultHandler
+
+    private val externalWebViewReqCode = 100
 
     companion object {
         private val appIdTag = "app_id_tag"
@@ -60,6 +69,9 @@ class MiniAppDisplayActivity : BaseActivity() {
 
         if (intent.hasExtra(miniAppTag) || intent.hasExtra(appIdTag)) {
             appId = intent.getStringExtra(appIdTag) ?: ""
+
+            if (appId.isEmpty())
+                appId = intent.getParcelableExtra<MiniAppInfo>(miniAppTag)!!.id
 
             setContentView(R.layout.mini_app_display_activity)
 
@@ -97,20 +109,40 @@ class MiniAppDisplayActivity : BaseActivity() {
                         AppPermission.getRequestCode(miniAppPermissionType)
                     )
                 }
+
+                override fun requestCustomPermissions(
+                    permissions: List<Pair<MiniAppCustomPermissionType, String>>,
+                    callback: (grantResult: String) -> Unit
+                ) {
+                    CustomPermissionPresenter().promptForCustomPermissions(
+                        this@MiniAppDisplayActivity,
+                        appId,
+                        permissions,
+                        callback
+                    )
+                }
             }
 
-            if (appId.isEmpty())
-                viewModel.obtainMiniAppDisplay(
-                    this@MiniAppDisplayActivity,
-                    intent.getParcelableExtra<MiniAppInfo>(miniAppTag)!!.id,
-                    miniAppMessageBridge
-                )
-            else
-                viewModel.obtainMiniAppDisplay(
-                    this@MiniAppDisplayActivity,
-                    appId,
-                    miniAppMessageBridge
-                )
+            miniAppNavigator = object : MiniAppNavigator {
+
+                override fun openExternalUrl(
+                    url: String,
+                    externalResultHandler: ExternalResultHandler
+                ) {
+                    sampleWebViewExternalResultHandler = externalResultHandler
+                    WebViewActivity.startForResult(
+                        this@MiniAppDisplayActivity, url,
+                        appId, externalWebViewReqCode
+                    )
+                }
+            }
+
+            viewModel.obtainMiniAppDisplay(
+                this@MiniAppDisplayActivity,
+                appId,
+                miniAppMessageBridge,
+                miniAppNavigator
+            )
         }
     }
 
@@ -121,6 +153,13 @@ class MiniAppDisplayActivity : BaseActivity() {
     ) {
         val isGranted = !grantResults.contains(PackageManager.PERMISSION_DENIED)
         miniappPermissionCallback.invoke(isGranted)
+    }
+
+    override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
+        super.onActivityResult(requestCode, resultCode, data)
+        if (requestCode == externalWebViewReqCode && resultCode == Activity.RESULT_OK) {
+            data?.let { intent -> sampleWebViewExternalResultHandler.emitResult(intent) }
+        }
     }
 
     private fun toggleProgressLoading(isOn: Boolean) {